Combine  VerticalPaymentMethodListView w/ ViewController
